First Commit
An Item is some kind of item.

An implemented item is an Item with a name, the individial price
and the quantity.

A CounterItem is the kind of item you would get at a counter, like the 
butcher or dairy counter, in which items are sold by weight.

The checkout is used for visual testing of printing the items,'
formatting the numbers etc.